requirements.txt
langchain==0.1.0
chromadb==0.4.18
pyyaml==6.0.1
python-dotenv==1.0.0
llama-index==0.9.3
transformers==4.36.0
torch==2.1.0pip
python-docx==0.8.11
PyPDF2==3.0.1
python-pptx==0.6.21
langchain_community.document_loaders